requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
idna==3.4
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
spotipy==2.22.1
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
